разделение вложенных data классов в dto
