+++ app/build.gradle
@@ -75,6 +75,7 @@ dependencies {
     implementation 'androidx.compose.material:material:1.4.3'
     implementation 'androidx.compose.material:material-icons-core:1.4.3'
     implementation 'androidx.compose.material:material-icons-extended:1.4.3'
+    implementation 'androidx.privacysandbox.tools:tools-core:1.0.0-alpha06'
 
     // Tests
     testImplementation 'junit:junit:4.13.2'
