+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonDto/Form.kt
@@ -1,0 +1,6 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonDto
+
+data class Form(
+    val name: String = "",
+    val url: String = ""
+)
