+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonDto/GameIndice.kt
@@ -1,0 +1,9 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonDto
+
+import com.google.gson.annotations.SerializedName
+
+data class GameIndice(
+    @SerializedName("game_index")
+    val gameIndex: Int = 0,
+    val version: Version = Version()
+)
