+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonDto/MoveLearnMethod.kt
@@ -1,0 +1,6 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonDto
+
+data class MoveLearnMethod(
+    val name: String = "",
+    val url: String = ""
+)
