+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonDto/Species.kt
@@ -1,0 +1,6 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonDto
+
+data class Species(
+    val name: String = "",
+    val url: String = ""
+)
