+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonDto/VersionGroup.kt
@@ -1,0 +1,6 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonDto
+
+data class VersionGroup(
+    val name: String = "",
+    val url: String = ""
+)
