+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonDto/VersionGroupDetail.kt
@@ -1,0 +1,12 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonDto
+
+import com.google.gson.annotations.SerializedName
+
+data class VersionGroupDetail(
+    @SerializedName("level_learned_at")
+    val levelLearnedAt: Int = 0,
+    @SerializedName("move_learn_method")
+    val moveLearnMethod: MoveLearnMethod = MoveLearnMethod(),
+    @SerializedName("version_group")
+    val versionGroup: VersionGroup = VersionGroup()
+)
