+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/pokemonList/Result.kt
@@ -1,0 +1,6 @@
+package com.biryulindevelop.pokemonlib.domain.dto.pokemonList
+
+data class Result(
+    val name: String = "",
+    val url: String = ""
+)
