+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/Animated.kt
@@ -1,0 +1,22 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class Animated(
+    @SerializedName("back_default")
+    val backDefault: String = "",
+    @SerializedName("back_female")
+    val backFemale: Any = Any(),
+    @SerializedName("back_shiny")
+    val backShiny: String = "",
+    @SerializedName("back_shiny_female")
+    val backShinyFemale: Any = Any(),
+    @SerializedName("front_default")
+    val frontDefault: String = "",
+    @SerializedName("front_female")
+    val frontFemale: Any = Any(),
+    @SerializedName("front_shiny")
+    val frontShiny: String = "",
+    @SerializedName("front_shiny_female")
+    val frontShinyFemale: Any = Any()
+)
