+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/DreamWorld.kt
@@ -1,0 +1,10 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class DreamWorld(
+    @SerializedName("front_default")
+    val frontDefault: String = "",
+    @SerializedName("front_female")
+    val frontFemale: Any = Any()
+)
