+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/GenerationI.kt
@@ -1,0 +1,9 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class GenerationI(
+    @SerializedName("red-blue")
+    val redBlue: RedBlue = RedBlue(),
+    val yellow: Yellow = Yellow()
+)
