+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/GenerationIi.kt
@@ -1,0 +1,7 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+data class GenerationIi(
+    val crystal: Crystal = Crystal(),
+    val gold: Gold = Gold(),
+    val silver: Silver = Silver()
+)
