+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/GenerationIv.kt
@@ -1,0 +1,11 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class GenerationIv(
+    @SerializedName("diamond-pearl")
+    val diamondPearl: DiamondPearl = DiamondPearl(),
+    @SerializedName("heartgold-soulsilver")
+    val heartgoldSoulsilver: HeartgoldSoulsilver = HeartgoldSoulsilver(),
+    val platinum: Platinum = Platinum()
+)
