+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/GenerationV.kt
@@ -1,0 +1,8 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class GenerationV(
+    @SerializedName("black-white")
+    val blackWhite: BlackWhite = BlackWhite()
+)
