+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/GenerationVi.kt
@@ -1,0 +1,10 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class GenerationVi(
+    @SerializedName("omegaruby-alphasapphire")
+    val omegarubyAlphasapphire: OmegarubyAlphasapphire = OmegarubyAlphasapphire(),
+    @SerializedName("x-y")
+    val xY: XY = XY()
+)
