+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/GenerationViii.kt
@@ -1,0 +1,5 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+data class GenerationViii(
+    val icons: Icons = Icons()
+)
