+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/Home.kt
@@ -1,0 +1,14 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class Home(
+    @SerializedName("front_default")
+    val frontDefault: String = "",
+    @SerializedName("front_female")
+    val frontFemale: Any = Any(),
+    @SerializedName("front_shiny")
+    val frontShiny: String = "",
+    @SerializedName("front_shiny_female")
+    val frontShinyFemale: Any = Any()
+)
