+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/OfficialArtwork.kt
@@ -1,0 +1,10 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class OfficialArtwork(
+    @SerializedName("front_default")
+    val frontDefault: String = "",
+    @SerializedName("front_shiny")
+    val frontShiny: String = ""
+)
