+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/RubySapphire.kt
@@ -1,0 +1,14 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class RubySapphire(
+    @SerializedName("back_default")
+    val backDefault: String = "",
+    @SerializedName("back_shiny")
+    val backShiny: String = "",
+    @SerializedName("front_default")
+    val frontDefault: String = "",
+    @SerializedName("front_shiny")
+    val frontShiny: String = ""
+)
