+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/Sprites.kt
@@ -1,0 +1,24 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class Sprites(
+    @SerializedName("back_default")
+    val backDefault: String = "",
+    @SerializedName("back_female")
+    val backFemale: Any = Any(),
+    @SerializedName("back_shiny")
+    val backShiny: String = "",
+    @SerializedName("back_shiny_female")
+    val backShinyFemale: Any = Any(),
+    @SerializedName("front_default")
+    val frontDefault: String = "",
+    @SerializedName("front_female")
+    val frontFemale: Any = Any(),
+    @SerializedName("front_shiny")
+    val frontShiny: String = "",
+    @SerializedName("front_shiny_female")
+    val frontShinyFemale: Any = Any(),
+    val other: Other = Other(),
+    val versions: Versions = Versions()
+)
