+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/domain/dto/sprites/Versions.kt
@@ -1,0 +1,22 @@
+package com.biryulindevelop.pokemonlib.domain.dto.sprites
+
+import com.google.gson.annotations.SerializedName
+
+data class Versions(
+    @SerializedName("generation-i")
+    val generationI: GenerationI = GenerationI(),
+    @SerializedName("generation-ii")
+    val generationIi: GenerationIi = GenerationIi(),
+    @SerializedName("generation-iii")
+    val generationIii: GenerationIii = GenerationIii(),
+    @SerializedName("generation-iv")
+    val generationIv: GenerationIv = GenerationIv(),
+    @SerializedName("generation-v")
+    val generationV: GenerationV = GenerationV(),
+    @SerializedName("generation-vi")
+    val generationVi: GenerationVi = GenerationVi(),
+    @SerializedName("generation-vii")
+    val generationVii: GenerationVii = GenerationVii(),
+    @SerializedName("generation-viii")
+    val generationViii: GenerationViii = GenerationViii()
+)
