+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/presentation/screens/details/PokemonDetailsViewModel.kt
@@ -2,7 +2,7 @@ package com.biryulindevelop.pokemonlib.presentation.screens.details
 
 import androidx.lifecycle.ViewModel
 import com.biryulindevelop.pokemonlib.data.repository.PokemonRepository
-import com.biryulindevelop.pokemonlib.domain.dto.PokemonDto
+import com.biryulindevelop.pokemonlib.domain.dto.pokemonDto.PokemonDto
 import com.biryulindevelop.pokemonlib.util.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
