+++ app/src/main/kotlin/com/biryulindevelop/pokemonlib/util/PokemonColorParse.kt
@@ -1,7 +1,8 @@
 package com.biryulindevelop.pokemonlib.util
 
 import androidx.compose.ui.graphics.Color
-import com.biryulindevelop.pokemonlib.domain.dto.PokemonDto
+import com.biryulindevelop.pokemonlib.domain.dto.pokemonDto.Stat
+import com.biryulindevelop.pokemonlib.domain.dto.pokemonDto.Type
 import com.biryulindevelop.pokemonlib.ui.theme.AtkColor
 import com.biryulindevelop.pokemonlib.ui.theme.DefColor
 import com.biryulindevelop.pokemonlib.ui.theme.HPColor
@@ -28,7 +29,7 @@ import com.biryulindevelop.pokemonlib.ui.theme.TypeRock
 import com.biryulindevelop.pokemonlib.ui.theme.TypeSteel
 import com.biryulindevelop.pokemonlib.ui.theme.TypeWater
 
-fun parseTypeToColor(type: PokemonDto.Type): Color {
+fun parseTypeToColor(type: Type): Color {
     return when (type.type.name.lowercase()) {
         "normal" -> TypeNormal
         "fire" -> TypeFire
@@ -52,7 +53,8 @@ fun parseTypeToColor(type: PokemonDto.Type): Color {
         else -> Color.Black
     }
 }
-fun changeTypeEngToRus(type: PokemonDto.Type): String {
+
+fun changeTypeEngToRus(type: Type): String {
     return when (type.type.name.lowercase()) {
         "normal" -> "Физический"
         "fire" -> "Огненный"
@@ -77,7 +79,7 @@ fun changeTypeEngToRus(type: PokemonDto.Type): String {
     }
 }
 
-fun parseStatColor(stat: PokemonDto.Stat): Color {
+fun parseStatColor(stat: Stat): Color {
     return when (stat.stat.name.lowercase()) {
         "hp" -> HPColor
         "attack" -> AtkColor
@@ -89,7 +91,7 @@ fun parseStatColor(stat: PokemonDto.Stat): Color {
     }
 }
 
-fun parseStatToAbbr(stat: PokemonDto.Stat): String {
+fun parseStatToAbbr(stat: Stat): String {
     return when (stat.stat.name.lowercase()) {
         "hp" -> "HP"
         "attack" -> "Atk"
