+++ app/src/main/res/drawable/poketball.xml
@@ -3,16 +3,16 @@
     android:height="1318.9dp"
     android:viewportWidth="1324.2"
     android:viewportHeight="1318.9">
-  <path
-      android:pathData="M685.6,0c5.5,1.3 11.1,0.8 16.7,1.1q33.2,2.1 66,7.4a644.7,644.7 0,0 1,90.9 21.3A663.1,663.1 0,0 1,1306.3 508.6a651.3,651.3 0,0 1,15 90.5c1.9,20.2 3.4,40.5 2.8,60.8 -0.7,23.9 -1.8,47.8 -4.5,71.6a659.2,659.2 0,0 1,-176.7 380,658.8 658.8,0 0,1 -224.4,155.7 645.6,645.6 0,0 1,-125.6 38.7A659.2,659.2 0,0 1,704.3 1317.6a673.9,673.9 0,0 1,-90 -0.3,659.3 659.3,0 0,1 -382.4,-157.2Q100.3,1046.9 40.1,884.1a636.9,636.9 0,0 1,-32.6 -127.7c-6.3,-42.4 -8.9,-85.1 -6.7,-128A662.3,662.3 0,0 1,523.6 14.6a676.6,676.6 0,0 1,92.2 -13c7.1,-0.5 14.1,-0.7 21.2,-1a26.5,26.5 0,0 0,2.9 -0.5Z"
-      android:fillColor="#383838"/>
-  <path
-      android:pathData="M1066.1,598.8q-107.6,0 -215.3,0c-2.3,0 -3.6,-0.3 -4.6,-2.9 -25.5,-61.9 -70.4,-101.6 -136,-115.6 -55.9,-12 -107.2,0.6 -152.5,35.5 -27.1,20.8 -46.7,47.5 -59,79.5 -1,2.6 -2.2,3.6 -5.2,3.6q-224.7,-0.1 -449.4,-0.1c-5,0 -5,0 -4.5,-4.9a575.6,575.6 0,0 1,12.8 -75.7A626.8,626.8 0,0 1,542.8 47.2c19.1,-3.7 38.4,-6.4 57.8,-8.3a620,620 0,0 1,68.4 -3A670,670 0,0 1,736.3 40.3,615 615,0 0,1 866.9,70.2a627.8,627.8 0,0 1,382 372.8,613.5 613.5,0 0,1 30.7,114.5c2.1,12.6 3.5,25.3 5.3,37.9 0.5,3.3 -1,3.5 -3.6,3.5Q1173.7,598.7 1066.1,598.8Z"
-      android:fillColor="#ee3f3e"/>
-  <path
-      android:pathData="M273.1,741.2q113.8,0 227.5,-0.1c2.4,0 3.6,0.7 4.7,2.8A186.4,186.4 0,0 0,674.8 848.4c38.5,-0.8 73.6,-12 105.1,-34.3A184.8,184.8 0,0 0,839.7 743.6a3.6,3.6 0,0 1,3.9 -2.4q217.9,0.1 435.8,0c4,0 2.6,1.8 2.4,4 -1.6,13.4 -4,26.8 -6.8,40a630.8,630.8 0,0 1,-23.6 83.3,623.1 623.1,0 0,1 -87.9,163.5 616.5,616.5 0,0 1,-72.8 81.5,625.3 625.3,0 0,1 -165.7,112 617.3,617.3 0,0 1,-122.1 41.8c-15.6,3.6 -31.3,6.7 -47.2,9.1 -11.9,1.8 -23.9,3 -35.9,4.2 -24.7,2.5 -49.5,2.9 -74.3,2.5 -13.5,-0.2 -27,-1.2 -40.5,-2.4 -10.7,-1 -21.4,-2.1 -32,-3.6A624,624 0,0 1,269.1 1144.6,623.8 623.8,0 0,1 195.3,1074.5c-56.1,-63.1 -98.3,-134.4 -125.7,-214.3a595.5,595.5 0,0 1,-22 -82c-2.1,-11.1 -3.8,-22.4 -5.8,-33.5 -0.5,-2.6 0.2,-3.5 3.1,-3.5Q159,741.3 273.1,741.2Z"
-      android:fillColor="#fdfcfc"/>
-  <path
-      android:pathData="M672.1,511.4c84.4,0.4 151.7,68.1 151.2,151.9a150.9,150.9 0,0 1,-301.8 -1.1C521.6,577.9 589.4,511.9 672.1,511.4Z"
-      android:fillColor="#fff"/>
+    <path
+        android:fillColor="#383838"
+        android:pathData="M685.6,0c5.5,1.3 11.1,0.8 16.7,1.1q33.2,2.1 66,7.4a644.7,644.7 0,0 1,90.9 21.3A663.1,663.1 0,0 1,1306.3 508.6a651.3,651.3 0,0 1,15 90.5c1.9,20.2 3.4,40.5 2.8,60.8 -0.7,23.9 -1.8,47.8 -4.5,71.6a659.2,659.2 0,0 1,-176.7 380,658.8 658.8,0 0,1 -224.4,155.7 645.6,645.6 0,0 1,-125.6 38.7A659.2,659.2 0,0 1,704.3 1317.6a673.9,673.9 0,0 1,-90 -0.3,659.3 659.3,0 0,1 -382.4,-157.2Q100.3,1046.9 40.1,884.1a636.9,636.9 0,0 1,-32.6 -127.7c-6.3,-42.4 -8.9,-85.1 -6.7,-128A662.3,662.3 0,0 1,523.6 14.6a676.6,676.6 0,0 1,92.2 -13c7.1,-0.5 14.1,-0.7 21.2,-1a26.5,26.5 0,0 0,2.9 -0.5Z" />
+    <path
+        android:fillColor="#ee3f3e"
+        android:pathData="M1066.1,598.8q-107.6,0 -215.3,0c-2.3,0 -3.6,-0.3 -4.6,-2.9 -25.5,-61.9 -70.4,-101.6 -136,-115.6 -55.9,-12 -107.2,0.6 -152.5,35.5 -27.1,20.8 -46.7,47.5 -59,79.5 -1,2.6 -2.2,3.6 -5.2,3.6q-224.7,-0.1 -449.4,-0.1c-5,0 -5,0 -4.5,-4.9a575.6,575.6 0,0 1,12.8 -75.7A626.8,626.8 0,0 1,542.8 47.2c19.1,-3.7 38.4,-6.4 57.8,-8.3a620,620 0,0 1,68.4 -3A670,670 0,0 1,736.3 40.3,615 615,0 0,1 866.9,70.2a627.8,627.8 0,0 1,382 372.8,613.5 613.5,0 0,1 30.7,114.5c2.1,12.6 3.5,25.3 5.3,37.9 0.5,3.3 -1,3.5 -3.6,3.5Q1173.7,598.7 1066.1,598.8Z" />
+    <path
+        android:fillColor="#fdfcfc"
+        android:pathData="M273.1,741.2q113.8,0 227.5,-0.1c2.4,0 3.6,0.7 4.7,2.8A186.4,186.4 0,0 0,674.8 848.4c38.5,-0.8 73.6,-12 105.1,-34.3A184.8,184.8 0,0 0,839.7 743.6a3.6,3.6 0,0 1,3.9 -2.4q217.9,0.1 435.8,0c4,0 2.6,1.8 2.4,4 -1.6,13.4 -4,26.8 -6.8,40a630.8,630.8 0,0 1,-23.6 83.3,623.1 623.1,0 0,1 -87.9,163.5 616.5,616.5 0,0 1,-72.8 81.5,625.3 625.3,0 0,1 -165.7,112 617.3,617.3 0,0 1,-122.1 41.8c-15.6,3.6 -31.3,6.7 -47.2,9.1 -11.9,1.8 -23.9,3 -35.9,4.2 -24.7,2.5 -49.5,2.9 -74.3,2.5 -13.5,-0.2 -27,-1.2 -40.5,-2.4 -10.7,-1 -21.4,-2.1 -32,-3.6A624,624 0,0 1,269.1 1144.6,623.8 623.8,0 0,1 195.3,1074.5c-56.1,-63.1 -98.3,-134.4 -125.7,-214.3a595.5,595.5 0,0 1,-22 -82c-2.1,-11.1 -3.8,-22.4 -5.8,-33.5 -0.5,-2.6 0.2,-3.5 3.1,-3.5Q159,741.3 273.1,741.2Z" />
+    <path
+        android:fillColor="#fff"
+        android:pathData="M672.1,511.4c84.4,0.4 151.7,68.1 151.2,151.9a150.9,150.9 0,0 1,-301.8 -1.1C521.6,577.9 589.4,511.9 672.1,511.4Z" />
 </vector>
